New upstream version 254~rc1

+++ test/units/testsuite-02.sh
@@ -3,13 +3,32 @@
 set -eux
 set -o pipefail
 
+if ! systemd-detect-virt -qc && [[ "${TEST_CMDLINE_NEWLINE:-}" != bar ]]; then
+    cat /proc/cmdline
+    echo >&2 "Expected TEST_CMDLINE_NEWLINE=bar from the kernel command line"
+    exit 1
+fi
+
+# If we're running with TEST_PREFER_NSPAWN=1 limit the set of tests we run
+# in QEMU to only those that can't run in a container to avoid running
+# the same tests again in a, most likely, very slow environment
+if ! systemd-detect-virt -qc && [[ "${TEST_PREFER_NSPAWN:-0}" -ne 0 ]]; then
+    TESTS_GLOB="test-loop-block"
+else
+    TESTS_GLOB=${TESTS_GLOB:-test-*}
+fi
+
 NPROC=$(nproc)
 MAX_QUEUE_SIZE=${NPROC:-2}
-TESTS_GLOB=${TESTS_GLOB:-test-*}
-mapfile -t TEST_LIST < <(find /usr/lib/systemd/tests/ -maxdepth 1 -type f -name "${TESTS_GLOB}")
+mapfile -t TEST_LIST < <(find /usr/lib/systemd/tests/unit-tests/ -maxdepth 1 -type f -name "${TESTS_GLOB}")
 
-# reset state
+# Reset state
 rm -fv /failed-tests /skipped-tests /skipped
+
+if ! systemd-detect-virt -qc; then
+    # Make sure ping works for unprivileged users (for test-bpf-firewall)
+    sysctl net.ipv4.ping_group_range="0 2147483647"
+fi
 
 # Check & report test results
 # Arguments:
@@ -24,7 +43,7 @@ function report_result() {
     local name="${1##*/}"
     local ret=$2
 
-    if [[ $ret -ne 0 && $ret != 77 ]]; then
+    if [[ $ret -ne 0 && $ret != 77 && $ret != 127 ]]; then
         echo "$name failed with $ret"
         echo "$name" >>/failed-tests
         {
@@ -32,7 +51,7 @@ function report_result() {
             cat "/$name.log"
             echo "--- $name end ---"
         } >>/failed
-    elif [[ $ret == 77 ]]; then
+    elif [[ $ret == 77 || $ret == 127 ]]; then
         echo "$name skipped"
         echo "$name" >>/skipped-tests
         {
@@ -60,7 +79,7 @@ for task in "${TEST_LIST[@]}"; do
             if ! kill -0 "${running[$key]}" &>/dev/null; then
                 # Task has finished, report its result and drop it from the queue
                 wait "${running[$key]}" && ec=0 || ec=$?
-                report_result "$key" $ec
+                report_result "$key" "$ec"
                 unset "running[$key]"
                 # Break from inner for loop and outer while loop to skip
                 # the sleep below when we find a free slot in the queue
@@ -83,8 +102,8 @@ done
 # Wait for remaining running tasks
 for key in "${!running[@]}"; do
     echo "Waiting for test '$key' to finish"
-    wait ${running[$key]} && ec=0 || ec=$?
-    report_result "$key" $ec
+    wait "${running[$key]}" && ec=0 || ec=$?
+    report_result "$key" "$ec"
     unset "running[$key]"
 done
